zemn.me: abstract hero video to own component.
