+++ project/zemn.me/app/BUILD.bazel
@@ -18,8 +18,6 @@ ts_project(
         "//:node_modules/next",
         "//:node_modules/react",
         "//project/zemn.me/app/experiments",
-        "//project/zemn.me/assets/kenwood",
-        "//project/zemn.me/assets/kenwood_snow",
         "//project/zemn.me/bio",
         "//project/zemn.me/components",
         "//project/zemn.me/components/TimeEye",
