+++ project/zemn.me/app/article/2014/csp/BUILD.bazel
@@ -2,8 +2,10 @@ load("//ts:rules.bzl", "ts_project")
 
 ts_project(
 	name = "csp",
+	assets = glob(["**/*.css"]),
 	deps = [
-		"//mdx:mdx_js"
+		"//mdx:mdx_js",
+		"//project/zemn.me/components/HeroVideo",
 	],
 	visibility = [ "//project/zemn.me:__subpackages__" ]
 )
