+++ project/zemn.me/app/article/2014/csp/page.tsx
@@ -1,4 +1,12 @@
 'use client';
-import { default as Page } from '#root/mdx/article/2014/csp';
+import { default as Content } from '#root/mdx/article/2014/csp';
+import { HeroVideo } from '#root/project/zemn.me/components/HeroVideo/hero_video.js';
+import style from '#root/project/zemn.me/app/article/2014/csp/style.module.css';
 
-export default Page;
+
+export default function Page() {
+	return <div className={style.article}>
+		<HeroVideo/>
+		<Content/>
+	</div>
+}
