+++ project/zemn.me/app/article/2014/csp/style.module.css
@@ -1,0 +1,3 @@
+.article {
+	font-size: 100%;
+}
