+++ project/zemn.me/app/style.module.css
@@ -52,7 +52,6 @@
 
 .headerBgv {
     z-index: -1;
-    object-fit: cover;
 }
 
 @media (orientation: landscape) {
