+++ project/zemn.me/components/HeroVideo/style.module.css
@@ -1,0 +1,3 @@
+.heroVideo {
+    object-fit: cover;
+}
